3-benchmark-llm/pyproject.toml
[project]
name = "3-benchmark-llm"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.0.0",
    "pydantic-ai-slim[openai]>=0.6.2",
    "python-dotenv>=1.1.1",
]
